# Optimise Chunk's dropRight & takeRight