--- a/core/shared/src/main/scala/zio/Chunk.scala
+++ b/core/shared/src/main/scala/zio/Chunk.scala
@@ -195,11 +195,23 @@ sealed abstract class Chunk[+A] extends ChunkLike[A] { self =>
     else if (n >= len) Chunk.empty
     else
       self match {
-        case Chunk.Slice(c, o, l)        => Chunk.Slice(c, o + n, l - n)
-        case Chunk.Singleton(_) if n > 0 => Chunk.empty
-        case c @ Chunk.Singleton(_)      => c
-        case Chunk.Empty                 => Chunk.empty
-        case _                           => Chunk.Slice(self, n, len - n)
+        case Chunk.Slice(c, o, l) => Chunk.Slice(c, o + n, l - n)
+        case _                    => Chunk.Slice(self, n, len - n)
+      }
+  }
+
+  /**
+   * Drops the last `n` elements of the chunk.
+   */
+  override def dropRight(n: Int): Chunk[A] = {
+    val len = self.length
+
+    if (n <= 0) self
+    else if (n >= len) Chunk.empty
+    else
+      self match {
+        case Chunk.Slice(c, o, l) => Chunk.Slice(c, o, l - n)
+        case _                    => Chunk.Slice(self, 0, len - n)
       }
   }
 
@@ -767,12 +779,20 @@ sealed abstract class Chunk[+A] extends ChunkLike[A] { self =>
     else if (n >= length) this
     else
       self match {
-        case Chunk.Empty => Chunk.Empty
-        case Chunk.Slice(c, o, l) =>
-          if (n >= l) this
-          else Chunk.Slice(c, o, n)
-        case c @ Chunk.Singleton(_) => c
-        case _                      => Chunk.Slice(self, 0, n)
+        case Chunk.Slice(c, o, l) => Chunk.Slice(c, o, n)
+        case _                    => Chunk.Slice(self, 0, n)
+      }
+
+  /**
+   * Takes the last `n` elements of the chunk.
+   */
+  override def takeRight(n: Int): Chunk[A] =
+    if (n <= 0) Chunk.Empty
+    else if (n >= length) this
+    else
+      self match {
+        case Chunk.Slice(c, o, l) => Chunk.Slice(c, o + l - n, n)
+        case _                    => Chunk.Slice(self, length - n, n)
       }
 
   /**