I removed some dead code from `Chunk#take` and `Chunk#drop` and added similar optimised overrides for `takeRight` and `dropRight`. Plus some tests & benchmarks.

Before (89f5141eed6a8c8eefce90f6bc82077ea041591c):
```
benchmarks/jmh:run -wi 3 -i 3 -f 2  zio.chunks.ChunkIndexedSeqComparison.drop* zio.chunks.ChunkIndexedSeqComparison.take*
[info] Benchmark                            (size)  Mode  Cnt     Score     Error  Units
[info] ChunkIndexedSeqComparison.drop         1000  avgt    6     4.779 ±   0.139  ns/op
[info] ChunkIndexedSeqComparison.dropRight    1000  avgt    6  2302.008 ±  39.221  ns/op
[info] ChunkIndexedSeqComparison.take         1000  avgt    6     4.552 ±   0.065  ns/op
[info] ChunkIndexedSeqComparison.takeRight    1000  avgt    6  2939.276 ± 187.680  ns/op
```

After (2f1b4089cd1b183a790bc1fd15e9418ec64bc385):
```
benchmarks/jmh:run -wi 3 -i 3 -f 2  zio.chunks.ChunkIndexedSeqComparison.drop* zio.chunks.ChunkIndexedSeqComparison.take*
[info] Benchmark                            (size)  Mode  Cnt  Score   Error  Units
[info] ChunkIndexedSeqComparison.drop         1000  avgt    6  4.566 ± 0.086  ns/op
[info] ChunkIndexedSeqComparison.dropRight    1000  avgt    6  4.566 ± 0.059  ns/op
[info] ChunkIndexedSeqComparison.take         1000  avgt    6  4.525 ± 0.048  ns/op
[info] ChunkIndexedSeqComparison.takeRight    1000  avgt    6  4.753 ± 0.095  ns/op